# api: adds CallWithStack to avoid allocations