--- a/api/wasm.go
+++ b/api/wasm.go
@@ -367,19 +367,15 @@ type Function interface {
 	Call(ctx context.Context, params ...uint64) ([]uint64, error)
 
 	// CallWithStack is an optimized variation of Call that saves memory
-	// allocations when the stack slice is large enough to hold the maximum
-	// length of function parameters and results.
+	// allocations when the stack slice is reused across calls.
 	//
+	// Stack length must be at least the max of parameter or result length.
 	// The caller adds parameters in order to the stack, and reads any results
-	// in order from the result stack, except in the error case.
+	// in order from the stack, except in the error case.
 	//
 	// # Notes
 	//
-	//   - When the stack parameter is large enough to hold the max count of
-	//     parameters and results, it is returned. Otherwise, a new slice is
-	//     allocated and returned. This means the caller should always prefer
-	//     using the result slice for subsequent calls.
-	//   - This is similar to GoFunc, except for using calling functions
+	//   - This is similar to GoModuleFunction, except for using calling functions
 	//     instead of implementing them. Moreover, this is used regardless of
 	//     whether the callee is a host or wasm defined function.
 	CallWithStack(ctx context.Context, stack []uint64) error