--- a/api/wasm.go
+++ b/api/wasm.go
@@ -373,6 +373,20 @@ type Function interface {
 	// The caller adds parameters in order to the stack, and reads any results
 	// in order from the stack, except in the error case.
 	//
+	// For example, the following reuses the same stack slice to call searchFn
+	// repeatedly saving one allocation per iteration:
+	//
+	//	stack := make([]uint64, 4)
+	//	for i, search := range searchParams {
+	//		// copy the next params to the stack
+	//		copy(stack, search)
+	//		if err := searchFn.CallWithStack(ctx, stack); err != nil {
+	//			return err
+	//		} else if stack[0] == 1 { // found
+	//			return i // searchParams[i] matched!
+	//		}
+	//	}
+	//
 	// # Notes
 	//
 	//   - This is similar to GoModuleFunction, except for using calling functions