--- a/internal/wasm/store_test.go
+++ b/internal/wasm/store_test.go
@@ -489,16 +489,15 @@ func (ce *mockCallEngine) Definition() api.FunctionDefinition { return nil }
 
 // Call implements the same method as documented on api.Function.
 func (ce *mockCallEngine) Call(ctx context.Context, _ ...uint64) (results []uint64, err error) {
-	return ce.CallWithStack(ctx, nil)
+	return nil, ce.CallWithStack(ctx, nil)
 }
 
 // CallWithStack implements the same method as documented on api.Function.
-func (ce *mockCallEngine) CallWithStack(_ context.Context, _ []uint64) (results []uint64, err error) {
+func (ce *mockCallEngine) CallWithStack(_ context.Context, _ []uint64) error {
 	if ce.callFailIndex >= 0 && ce.index == Index(ce.callFailIndex) {
-		err = errors.New("call failed")
-		return
+		return errors.New("call failed")
 	}
-	return
+	return nil
 }
 
 func TestStore_getFunctionTypeID(t *testing.T) {