This implements the API discussed in #1332.

It still needs unit tests, and redoing the benchmarks to confirm the improvements.