@ncruces 2023-04-26
Re-ran the benchmarks from #1332 ([`github.com/maragudk/sqlite-benchmark`](https://github.com/maragudk/sqlite-benchmark) adapted to use my [driver](https://pkg.go.dev/github.com/ncruces/go-sqlite3/driver) rather than [mattn](https://pkg.go.dev/github.com/mattn/go-sqlite3)).

Again this benchmark is I/O bound, so improvements to CPU time are negligible, and there's plenty of non-determinism. OTOH, it's not a synthetic made to make this look great.

The reduction in `allocs/op` is stark (this is a rough proxy for the amount of GC work, not necessarily peak memory usage).

Before:
```
goos: darwin
goarch: amd64
pkg: github.com/maragudk/sqlite-benchmark
cpu: Intel(R) Core(TM) i7-8750H CPU @ 2.20GHz
BenchmarkSelect1-12                                         2000              8643 ns/op           28301 B/op         13 allocs/op
BenchmarkDB_ReadPost/mutex_false-12                         2000             30699 ns/op           26302 B/op         76 allocs/op
BenchmarkDB_ReadPost/mutex_true-12                          2000             29405 ns/op           28953 B/op         77 allocs/op
BenchmarkDB_WritePost/mutex_false-12                        2000            982031 ns/op           20309 B/op         37 allocs/op
BenchmarkDB_WritePost/mutex_true-12                         2000            767473 ns/op            7634 B/op         36 allocs/op
BenchmarkDB_ReadPostAndMaybeWriteComment/mutex_false/comment_rate_0.01-12                   2000             43738 ns/op           24638 B/op        218 allocs/op
BenchmarkDB_ReadPostAndMaybeWriteComment/mutex_false/comment_rate_0.1-12                    2000            195674 ns/op           31664 B/op       1476 allocs/op
BenchmarkDB_ReadPostAndMaybeWriteComment/mutex_false/comment_rate_1-12                      2000          11684738 ns/op          230575 B/op      15605 allocs/op
BenchmarkDB_ReadPostAndMaybeWriteComment/mutex_true/comment_rate_0.01-12                    2000             62366 ns/op           96816 B/op        220 allocs/op
BenchmarkDB_ReadPostAndMaybeWriteComment/mutex_true/comment_rate_0.1-12                     2000            301926 ns/op          627289 B/op       1633 allocs/op
BenchmarkDB_ReadPostAndMaybeWriteComment/mutex_true/comment_rate_1-12                       2000           2598808 ns/op          225410 B/op      15603 allocs/op
BenchmarkDB_ReadPostAndMaybeWriteCommentWithPool/comment_rate_0.01-12                       2000             42212 ns/op           36322 B/op        183 allocs/op
BenchmarkDB_ReadPostAndMaybeWriteCommentWithPool/comment_rate_0.1-12                        2000            202609 ns/op          102164 B/op       1727 allocs/op
BenchmarkDB_ReadPostAndMaybeWriteCommentWithPool/comment_rate_1-12                          2000           2937426 ns/op          225967 B/op      15606 allocs/op
PASS
ok      github.com/maragudk/sqlite-benchmark    40.663s
```

After:
```
goos: darwin
goarch: amd64
pkg: github.com/maragudk/sqlite-benchmark
cpu: Intel(R) Core(TM) i7-8750H CPU @ 2.20GHz
BenchmarkSelect1-12                                         2000             10291 ns/op           23294 B/op          8 allocs/op
BenchmarkDB_ReadPost/mutex_false-12                         2000             29984 ns/op           26373 B/op         47 allocs/op
BenchmarkDB_ReadPost/mutex_true-12                          2000             33873 ns/op           32017 B/op         49 allocs/op
BenchmarkDB_WritePost/mutex_false-12                        2000            973562 ns/op           24745 B/op         24 allocs/op
BenchmarkDB_WritePost/mutex_true-12                         2000            794107 ns/op            7523 B/op         23 allocs/op
BenchmarkDB_ReadPostAndMaybeWriteComment/mutex_false/comment_rate_0.01-12                   2000             38152 ns/op           29614 B/op         74 allocs/op
BenchmarkDB_ReadPostAndMaybeWriteComment/mutex_false/comment_rate_0.1-12                    2000            320714 ns/op           22022 B/op        509 allocs/op
BenchmarkDB_ReadPostAndMaybeWriteComment/mutex_false/comment_rate_1-12                      2000          10405274 ns/op          159338 B/op       6665 allocs/op
BenchmarkDB_ReadPostAndMaybeWriteComment/mutex_true/comment_rate_0.01-12                    2000             68912 ns/op          100010 B/op        117 allocs/op
BenchmarkDB_ReadPostAndMaybeWriteComment/mutex_true/comment_rate_0.1-12                     2000            315064 ns/op          617469 B/op        758 allocs/op
BenchmarkDB_ReadPostAndMaybeWriteComment/mutex_true/comment_rate_1-12                       2000           2373769 ns/op          154330 B/op       6664 allocs/op
BenchmarkDB_ReadPostAndMaybeWriteCommentWithPool/comment_rate_0.01-12                       2000             53201 ns/op           46634 B/op        123 allocs/op
BenchmarkDB_ReadPostAndMaybeWriteCommentWithPool/comment_rate_0.1-12                        2000            192710 ns/op           96767 B/op        662 allocs/op
BenchmarkDB_ReadPostAndMaybeWriteCommentWithPool/comment_rate_1-12                          2000           2552369 ns/op          155078 B/op       6666 allocs/op
PASS
ok      github.com/maragudk/sqlite-benchmark    37.593s
```