@ncruces 2023-04-26
@inkeliz, since you wanted this, I'd really appreciate your feedback on whether this solves your issue, and what's the more ergonomic API.

I don't want to get too much caught up into my own solution.

Thanks!