@ncruces 2023-04-26
I think in #1332 and here we've established the usefulness of this, but can't settle for an API design (which in this case, I find, is understandably hard).

I'm happy to keep this updated (tests passing) until more people are ready to share instances where they'd use this, so we can improve ergonomics and settle on a design.

Or someone makes the call, because the perfect is the enemy of the good.

I'd use this in exactly one call site, so I'm probably not the right one to make the call, any design works for me.