@codefromthecrypt 2023-04-27
My summary comment is that the api surface of a codebase should have coherence with itself as a feature, even if it wasn't also optimizable due to this.

It would almost be self-deprecating humor to have to tell people the secret way to make our API whose goal is to reduce allocations, do less via things like this: `CallWithStack(ctx, params, params[:0])` when we already have an API design the same people will have to understand to reduce allocations host side.

When designing something explicitly for advanced users, we shouldn't fear them being able to do min(len(params), len(results)) if trying to size things. The exact concept is required for the host side, even if they aren't the ones sizing the slice.

All this said, If others feel like they really want the third API style and that's better than two, go for it. I've said my piece and feel folks here can grok where I'm coming from.