@codefromthecrypt 2023-04-27
If we do go for api style 3, please don't link it to GoModuleFunc, as it no longer has any relationship to it.